% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{columns.data.frame}
\alias{columns.data.frame}
\title{Get Data Frame Columns Except those Specified by Name}
\usage{
columns.data.frame(df, ex)
}
\arguments{
\item{df}{A data frame from which column names will be retrieved.}

\item{ex}{A character vector of column names to exclude.}
}
\value{
A data frame with columns  excluding those in \code{ex}.
}
\description{
This function returns the columns of a data frame excluding those specified.
}
\examples{
df <- data.frame(a = 1:3, b = 4:6, c = 7:9)
columns.data.frame(df, c("b", "c"))

}
