% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{diag.plot.lme}
\alias{diag.plot.lme}
\title{Diagnostic Plot for Linear Mixed-Effects Model}
\usage{
diag.plot.lme(model)
}
\arguments{
\item{model}{An object of class \code{lme}, representing a linear
mixed-effects model.}
}
\value{
A \code{ggarrange} object containing three diagnostic plots:
\itemize{
\item A: Residuals vs Fitted values plot.
\item B: Q-Q plot of the residuals.
\item C: Histogram of the residuals.
}
}
\description{
This function generates diagnostic plots for a linear mixed-effects
model (\code{lme}).
The function creates three plots: a residuals vs fitted values plot,
a Q-Q plot of the residuals,
and a histogram of the residuals. These plots are arranged in a single
figure to visually assess the model's performance.
}
\details{
The function takes a fitted \code{lme} model object and uses \code{ggplot2} and
related functions to create three diagnostic plots.
}
\examples{
\dontrun{
library(nlme)
library(ggplot2)
model <- lme(
  fixed = distance ~ age, random = ~ 1 | Subject,
  data = Orthodont
)
diag.plot.lme(model)
}

}
