% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{diag.plot.lmer}
\alias{diag.plot.lmer}
\title{Diagonal plot for linear mixed effects models}
\usage{
diag.plot.lmer(model)
}
\arguments{
\item{model}{a linear mixed effects model created using the \code{lme4} package}
}
\value{
a \code{ggplot2} object representing the diagonal plot
}
\description{
This function creates a diagonal plot for a linear mixed effects model using
the \code{ggplot2} package. The plot includes a scatterplot of the residuals
versus the fitted values, a Q-Q plot of the residuals, and a histogram of the
residuals.
}
\examples{
model <- lmer(Sepal.Length ~ Sepal.Width + (1 | Species), data = iris)
diag.plot.lmer(model)

}
