% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{make_factor}
\alias{make_factor}
\title{Create a Factor with Custom Levels and Labels}
\usage{
make_factor(input_vector, levels, labels, explicit_na = TRUE)
}
\arguments{
\item{input_vector}{A vector to be converted into a factor.}

\item{levels}{A vector of levels to be assigned to the factor.}

\item{labels}{A vector of labels corresponding to the levels.}

\item{explicit_na}{Logical, indicating whether missing values (NA) should
be explicitly assigned to a "Missing" level. Default is TRUE.}
}
\value{
A factor with specified levels and labels, potentially including
an explicit "Missing" level for NA values.
}
\description{
This function transforms a vector into a factor, assigns specified levels
and labels, and optionally handles missing values by explicitly assigning
them a level.
}
\details{
This function allows for the creation of factors from a vector while
specifying both levels and labels.
If \code{explicit_na} is set to \code{TRUE}, missing values are explicitly
assigned to a level named "Missing".
The function also drops any unused levels from the factor.
}
\examples{
vec <- c("low", "medium", "high", NA, "low")
levels <- c("low", "medium", "high")
labels <- c("Low", "Medium", "High")
make_factor(vec, levels, labels)

}
