% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{names_of_df_except}
\alias{names_of_df_except}
\title{Get Data Frame Column Names Except Specified}
\usage{
names_of_df_except(df, ex)
}
\arguments{
\item{df}{A data frame from which column names will be retrieved.}

\item{ex}{A character vector of column names to exclude.}
}
\value{
A character vector of column names from the data frame, excluding those in \code{ex}.
}
\description{
This function returns the column names of a data frame excluding those specified.
}
\examples{
df <- data.frame(a = 1:3, b = 4:6, c = 7:9)
names_of_df_except(df, c("b", "c"))

}
