% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{present.factor}
\alias{present.factor}
\title{Present Factor Variable as a Bar Plot and Summary Table}
\usage{
present.factor(Datasource, var.name)
}
\arguments{
\item{Datasource}{A data frame containing the factor variable to be analyzed.}

\item{var.name}{A string representing the name of the factor variable to be
plotted and summarized.}
}
\value{
A list with two elements:
\item{graph}{A ggplot object representing the bar plot of the factor
variable.}
\item{table}{A data frame with two columns: \code{N} (the count of each factor
level) and \code{Percentage} (the percentage of total observations for each
level).}
}
\description{
This function generates a bar plot and a summary table for a factor variable
from a given data source. The bar plot displays the counts of each level, and
the summary table includes both counts and percentages of each level.
}
\details{
This function takes a data frame and a factor variable name, and generates
both a bar plot (using \code{ggplot2}) and a summary table that shows the counts
and percentages of each level of the factor.
}
\examples{
\dontrun{
library(ggplot2)
library(dplyr)
data(iris)
result <- present.factor(iris, "Species")
result$graph # Display the bar plot
result$table # Display the summary table
}

}
