% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{rlmerMod.to.glm}
\alias{rlmerMod.to.glm}
\title{Convert a \code{rlmerMod} Model to a \code{glm} Model}
\usage{
rlmerMod.to.glm(mod)
}
\arguments{
\item{mod}{A fitted \code{rlmerMod} model object from the \code{rstanarm} package.}
}
\value{
A \code{glm} model object with the following attributes:
\itemize{
\item \code{coefficients}: Fixed effects coefficients from the original
model.
\item \code{vcov}: Variance-covariance matrix of the fixed effects
coefficients.
\item \code{linear.predictors}: Linear predictors for the \code{glm} model.
\item \code{fitted.values}: Fitted values based on the \code{glm} model.
\item \code{weights}: Weights used in the \code{glm} model.
\item \code{residuals}: Residuals calculated for the \code{glm} model.
}
The object is assigned a new class \code{fakeglm} in addition to the standard
\code{glm} class.
}
\description{
This function converts a fitted \code{rlmerMod} model object (from the \code{rstanarm}
package) to a generalized linear model (\code{glm}) object. This is done by
extracting the fixed effects and variance-covariance matrix from the
\code{rlmerMod} model and creating a new \code{glm} model with equivalent fixed
effects.
}
\details{
The function extracts the call, fixed effects, and
variance-covariance matrix from the \code{rlmerMod} model. It then constructs
a new \code{glm} object with these components and adjusts its attributes to
match the original model's structure.
}
\examples{
library(rstanarm)
# Fit a linear mixed-effects model
mod <- rlmer(Reaction ~ Days + (Days | Subject), data = sleepstudy)
# Convert the model to a glm object
mod_glm <- rlmerMod.to.glm(mod)
# Examine the glm object
summary(mod_glm)
}
