% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{separate_baseline_df}
\alias{separate_baseline_df}
\title{Separate Baseline Scores in a Data Frame}
\usage{
separate_baseline_df(
  df,
  time_var = "Time",
  baseline_lvl = "Baseline",
  var_name = "Outcome",
  id_vars = c("PartId", "Group")
)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{time_var}{A string specifying the name of the time variable (default is "Time").}

\item{baseline_lvl}{A string specifying the baseline level (default is "Baseline").}

\item{var_name}{A string specifying the name of the outcome variable (default is "Outcome").}

\item{id_vars}{A character vector of identifier variables (default is c("PartId", "Group")).}
}
\value{
A data frame that includes the non-baseline records and adds a column for the baseline outcome.
}
\description{
This function separates baseline data in a data frame by separating baseline and non-baseline records.
}
\examples{
df <- data.frame(PartId = c(1, 1, 2, 2),
                 Group = c("A", "A", "B", "B"),
                 Time = c("Baseline", "Post", "Baseline", "Post"),
                 Outcome = c(10, 12, 20, 22))

result <- separate_baseline_df(df)
print(result)

}
