% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{standard.em.table}
\alias{standard.em.table}
\title{Create a standard format table from raw estimated marginal means table}
\usage{
standard.em.table(raw.em.table)
}
\arguments{
\item{raw.em.table}{A data frame containing raw estimated marginal means and
corresponding standard errors.}
}
\value{
A data frame in a standardized format with estimated means, standard errors,
and confidence intervals. The confidence intervals are calculated based on the
\code{ci.names} parameter, which defaults to "lower.CL" and "upper.CL".
}
\description{
This function takes a raw estimated marginal means (EMM) table, calculated using
a statistical model, and transforms it into a standardized format with estimated
means, standard errors, and confidence intervals.
}
\examples{
library(emmeans)
data("airquality")
model <- lm(Ozone ~ WindTemp, data = airquality)
em.table <- emmeans::emmeans(model, ~WindTemp)
standard.em.table(em.table)

}
