% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{std_beta_lmer}
\alias{std_beta_lmer}
\title{Calculate Standardized Beta Coefficients from a Linear Mixed Model}
\usage{
std_beta_lmer(mod)
}
\arguments{
\item{mod}{A fitted linear mixed-effects model object of class \code{lmerMod},
typically created with the \code{lmer} function from the \code{lme4} package.}
}
\value{
A vector of standardized beta coefficients corresponding to the
fixed effects (excluding the intercept) in the model.
}
\description{
This function calculates standardized beta coefficients from a fitted linear
mixed-effects model (using the \code{lmer} function from the \code{lme4} package).
Standardized beta coefficients are obtained by scaling the fixed effects
coefficients by the standard deviations of the predictor variables and
the outcome variable.
}
\examples{
library(lme4)
# Fit a linear mixed-effects model
mod <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
# Calculate standardized beta coefficients
std_beta_lmer(mod)
}
