% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{var.table.lmer}
\alias{var.table.lmer}
\title{Variance Table for Linear Mixed-Effects Model}
\usage{
var.table.lmer(model)
}
\arguments{
\item{model}{An object of class \code{lmerMod}, representing a linear
mixed-effects model, typically fitted using the \code{lme4} package.}
}
\value{
A data frame with the following columns:
\item{Component}{The type of variance component: "Fixed", "Random",
or "Error".}
\item{Variance}{The amount of variance attributable to each component.}
\item{Percentage}{The percentage of total variance explained by each
component.}
\item{ICC}{The Intraclass Correlation Coefficient (ICC) for the random
component, if applicable.}
}
\description{
This function calculates and returns a table of variance components
for a linear mixed-effects model (\code{lmer}).
The table includes the fixed, random, and error components, as well as
their percentages relative to the total variance
and the Intraclass Correlation Coefficient (ICC).
}
\details{
The function extracts variance components using the
\code{insight::get_variance()} function and calculates the percentage
of variance explained by each component. Additionally, the ICC for the
random component is calculated.
}
\examples{
\dontrun{
library(lme4)
library(insight)
model <- lmer(Reaction ~ Days + (Days | Subject), data = sleepstudy)
var.table.lmer(model)
}

}
